Tweaks to ColouredChord
Changed the print in __repr__ to return rather than print to fix the error being raised.

Removed the check for raw_chord starting with "[" as this will no longer be needed. It will be in the new Class "Part"

Also made tweaks to reword_unclean_chord_name. Added in handling of m7b5 (gets changed to min7b5) and also fixed a bug in the code that was missed.