Refactor to get Song to do cleanup
Two main changes
1. Clean up is done within class initialising
2. Coloured image generation is within Song rather than it's own function

Also when you request a url from get_unclean_chord_data it returns the unclean song structured nicely in a dictionary e.g. {'intro' : ['Cmaj', 'D', 'Emin']}

So this can be fed into Song() and the clean up happens within the initialising.

This is nice to allow quick pulling of songs and clean up is done with less steps.