Reworked major key dictionary
The dictionary containing the major keys' chords now has all enharmonic equivalent chords in it e.g. a C#maj chord is also a Dbmaj chord.

This means that regardless of however the chords of input e..g if someone uses all flats or all sharps, the correct major key will be returned